refactor: file name style and move implementation
- user lower-case, single-word file name
- move all implementations that depend VSCode API to extension.ts
