+++ src/extension.ts
@@ -4,10 +4,9 @@ import { tmpdir } from "os";
 import * as path from "path";
 
 import * as gpg from './indicator/gpg';
-import { VscodeOutputLogger } from './logger';
-import SecretObjectStorage from "./ObjectStorages/SecretObjectStorage";
-import MementoObjectStorage from "./ObjectStorages/MementoObjectStorage";
-import KeyStatusManager from "./KeyStatusManager";
+import { Logger } from "./indicator/logger";
+import KeyStatusManager from "./manager";
+import { AbstractObjectStorage, StorageObject } from "./manager";
 import { m } from "./message";
 
 type statusStyleEnum = "fingerprintWithUserId" | "fingerprint" | "userId";
@@ -315,4 +314,101 @@ export async function deactivate() {
         recursive: true,
         force: true,
     });
+}
+
+const timeStr = (date = new Date()) => date.toISOString();
+
+enum LogLevel {
+    error = 1,
+    warning,
+    info
+}
+
+export class VscodeOutputLogger implements Logger {
+    private outputChannel: vscode.OutputChannel;
+    private level: LogLevel;
+    /**
+     * @param name - The name of VS Code output channel on UI
+     * @param level - The log level for the logger
+     */
+    constructor(name: string, level: string) {
+        this.outputChannel = vscode.window.createOutputChannel(name);
+        this.level = VscodeOutputLogger.levelFromString(level);
+    }
+
+    static levelFromString(level: string): LogLevel {
+        switch (level) {
+            case "error":
+                return LogLevel.error;
+            case "warning":
+                return LogLevel.warning;
+            case "info":
+                return LogLevel.info;
+            default:
+                throw new Error(`unknown log level: ${level}`);
+        }
+    }
+
+    setLevel(level: string): void {
+        this.level = VscodeOutputLogger.levelFromString(level);
+    }
+
+    /**
+     * Log some message at info level.
+     * @param message - a message without ending new line
+     */
+    info(message: string): void {
+        if (this.level < LogLevel.info) {
+            return;
+        }
+        this.outputChannel.appendLine(`[${timeStr()}] [INFO] ` + message);
+    }
+
+    /**
+     * Log some message at warning level.
+     * @param message - a message without ending new line
+     */
+    warn(message: string): void {
+        if (this.level < LogLevel.warning) {
+            return;
+        }
+        this.outputChannel.appendLine(`[${timeStr()}] [WARN] ` + message);
+    }
+
+    /**
+     * Log some message at error level.
+     * @param message - a message without ending new line
+     */
+    error(message: string): void {
+        if (this.level < LogLevel.error) {
+            return;
+        }
+        this.outputChannel.appendLine(`[${timeStr()}] [ERROR] ` + message);
+    }
+}
+
+
+class MementoObjectStorage extends AbstractObjectStorage {
+    constructor(private storageUtility: vscode.Memento, protected logger: VscodeOutputLogger) {
+        super();
+    }
+    protected async _getAll(): Promise<StorageObject> {
+        return await this.storageUtility.get(this.key) || {};
+    }
+    protected async _setStorage(storage: StorageObject): Promise<void> {
+        await this.storageUtility.update(this.key, storage);
+    }
+}
+
+
+class SecretObjectStorage extends AbstractObjectStorage {
+    constructor(private storageUtility: vscode.SecretStorage, protected logger: VscodeOutputLogger) {
+        super();
+    }
+    protected async _getAll(): Promise<StorageObject> {
+        return JSON.parse(await this.storageUtility.get(this.key) || "{}");
+    }
+    protected async _setStorage(storage: StorageObject): Promise<void> {
+        await this.storageUtility.store(this.key, JSON.stringify(storage));
+    }
 }
